oauth/onboard/agents: collapse handlers into the matching unit file
web.go was a grab-bag of API handlers across every dashboard surface.
Move the OAuth, onboard, and agents handlers next to the registry
that backs them — fewer files, each file owns one logical unit.

oauth.go now carries (in addition to OAuthRegistry / refresh /
persistence): apiOAuthStart / Exchange / DevicePoll / Revoke,
startDeviceFlow, pollDeviceFlow, exchangeOAuthCode + the Anthropic
JSON-body variant, plus the oauthSession type used by them.

onboard.go now carries: apiOnboardStart / Approve / Claim / Poll /
Lookup alongside the onboardRegistry that owns the lifecycle.

agents.go now carries: apiAgents / AgentDelete / AgentProfile /
Profiles / Status, RuleSummary + collectRuleSummaries +
matchSourceMap (read off agents' profile), plus IntegrationRow /
Owner / credentialsInProfile.

web.go shrinks to dashboard auth + the mux + a few cross-cut
handlers. No behaviour change.

  before:  web.go  1895
  after:   web.go  1048   oauth.go +325  onboard.go +200  agents.go +320
