+++ web.go
@@ -2,12 +2,10 @@ package main
 
 import (
 	"bytes"
-	"context"
 	"crypto/sha256"
 	"crypto/subtle"
 	"database/sql"
 	"embed"
-	"encoding/base64"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
@@ -15,18 +13,13 @@ import (
 	"html/template"
 	"io"
 	"io/fs"
-	"log"
-	"net"
 	"net/http"
 	"net/url"
 	"os"
-	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
-
-	"golang.org/x/oauth2"
 
 	"github.com/denoland/clawpatrol/config"
 	"github.com/denoland/clawpatrol/config/runtime"
@@ -39,30 +32,6 @@ var dashboardFS embed.FS
 var loginHTML string
 
 var loginTpl = template.Must(template.New("login").Parse(loginHTML))
-
-type IntegrationRow struct {
-	ID       string              `json:"id"`
-	Name     string              `json:"name"`
-	Type     string              `json:"type"` // credential plugin type
-	HasOAuth bool                `json:"has_oauth"`
-	Slots    []config.SecretSlot `json:"slots,omitempty"`
-	Owners   []Owner             `json:"owners"`
-}
-
-type Owner struct {
-	Owner     string `json:"owner"`
-	Connected bool   `json:"connected"`
-	ExpiresAt int64  `json:"expires_at,omitempty"`
-}
-
-type oauthSession struct {
-	verifier string
-	state    string
-	cfg      *oauth2.Config
-	id       string
-	owner    string
-	created  time.Time
-}
 
 type webMux struct {
 	g         *Gateway
@@ -445,76 +414,6 @@ func (w *webMux) apiWhoami(rw http.ResponseWriter, r *http.Request) {
 // by profile when ?profile=NAME is set — only credentials referenced
 // by an endpoint in that profile come back. Without the param, every
 // declared credential ships (root view).
-func (w *webMux) apiStatus(rw http.ResponseWriter, r *http.Request) {
-	out := []IntegrationRow{}
-	policy := w.g.policy.Load()
-	if policy == nil {
-		writeJSON(rw, out)
-		return
-	}
-	profile := r.URL.Query().Get("profile")
-	allowed := credentialsInProfile(policy, profile) // nil = no filter
-
-	names := make([]string, 0, len(policy.Credentials))
-	for name := range policy.Credentials {
-		if allowed != nil && !allowed[name] {
-			continue
-		}
-		names = append(names, name)
-	}
-	sort.Strings(names)
-	caller, _ := w.ownerForCaller(r)
-	for _, name := range names {
-		ent := policy.Credentials[name]
-		row := IntegrationRow{ID: name, Name: name, Type: ent.Plugin.Type}
-		if _, ok := ent.Body.(config.OAuthFlowProvider); ok {
-			row.HasOAuth = true
-			for _, owner := range w.g.oauth.Owners(name) {
-				connected, exp := w.g.oauth.Status(name, owner)
-				o := Owner{Owner: owner, Connected: connected}
-				if !exp.IsZero() {
-					o.ExpiresAt = exp.Unix()
-				}
-				row.Owners = append(row.Owners, o)
-			}
-		}
-		if sp, ok := ent.Body.(config.SecretSlotsProvider); ok {
-			row.Slots = sp.SecretSlots()
-			if caller != "" {
-				present, _ := credentialSlotPresence(w.g.db, name, caller)
-				if len(present) > 0 {
-					row.Owners = append(row.Owners, Owner{Owner: caller, Connected: true})
-				}
-			}
-		}
-		out = append(out, row)
-	}
-	writeJSON(rw, out)
-}
-
-// credentialsInProfile returns the set of credential bare names that
-// any endpoint in the given profile references. nil means "no filter
-// — return everything." Used by apiStatus and the device-page card
-// render so per-device views only show credentials the device's
-// profile actually uses.
-func credentialsInProfile(policy *config.CompiledPolicy, profile string) map[string]bool {
-	if profile == "" || policy == nil {
-		return nil
-	}
-	prof, ok := policy.Profiles[profile]
-	if !ok {
-		return map[string]bool{} // unknown profile → empty set, not nil
-	}
-	out := map[string]bool{}
-	for _, ep := range prof.Endpoints {
-		for _, cb := range ep.Credentials {
-			if cb.Credential != nil {
-				out[cb.Credential.Symbol.Name] = true
-			}
-		}
-	}
-	return out
-}
 
 // apiCredentialsSet persists one or more slot values for a non-OAuth
 // credential. Owner defaults to the caller's profile. Body shape:
@@ -634,236 +533,6 @@ func lookupOAuthFlow(policy *config.CompiledPolicy, name string) *config.OAuthIn
 	return fp.OAuthFlow()
 }
 
-func (w *webMux) apiAgents(rw http.ResponseWriter, _ *http.Request) {
-	var snap []*Agent
-	if w.g.agents != nil {
-		snap = w.g.agents.snapshot()
-	}
-	// External IPs: the underlay v4/v6 each WG peer is dialing in from.
-	// Show these in place of the server-side /32 (routing artefact).
-	// Live endpoint observed via wg-go IpcGet — persist into the devices
-	// row so the dashboard keeps a stable last-known address even when
-	// the peer goes idle and wg-go drops its endpoint state.
-	if globalWG != nil {
-		for ip, ep := range globalWG.EndpointsByIP() {
-			if ep == "" {
-				continue
-			}
-			parsed := net.ParseIP(ep)
-			var v4, v6 string
-			if p4 := parsed.To4(); p4 != nil {
-				v4 = p4.String()
-			} else if parsed != nil {
-				v6 = parsed.String()
-			}
-			w.onboard.SetExternalIPs(ip, v4, v6)
-		}
-	}
-	for _, a := range snap {
-		v4, v6 := w.onboard.ExternalIPs(a.IP)
-		a.ExternalIPv4, a.ExternalIPv6 = v4, v6
-	}
-	// enrich with connected integrations per agent's user. Re-run on
-	// every snapshot — caching by `Integrations != nil` would freeze
-	// the list at first sighting (a freshly-onboarded device whose
-	// user later connects Claude wouldn't reflect the new connection).
-	for _, a := range snap {
-		// Per-profile credentials: the agent's Integrations list
-		// reflects what the device's bound profile has connected. Falls
-		// back to the legacy per-user lookup for tailnet-control-mode
-		// installs that still bucket creds by login.
-		profile := w.onboard.ProfileForIP(a.IP)
-		if profile == "" {
-			if a.User == "" || a.User == "tagged-devices" {
-				if owner := w.onboard.OwnerForIP(a.IP); owner != "" {
-					a.User = owner
-				}
-			}
-			profile = a.User
-		}
-		if profile == "" {
-			continue
-		}
-		// Walk every declared credential — connected if either an
-		// OAuth token exists OR the operator pasted a secret slot via
-		// the dashboard. Was hardcoded to the claude/codex/github
-		// legacy trio, which silently hid every other credential type
-		// from the agents table.
-		var ids []string
-		if policy := w.g.Policy(); policy != nil {
-			names := make([]string, 0, len(policy.Credentials))
-			for name := range policy.Credentials {
-				names = append(names, name)
-			}
-			sort.Strings(names)
-			for _, name := range names {
-				if conn, _ := w.g.oauth.Status(name, profile); conn {
-					ids = append(ids, name)
-					continue
-				}
-				present, _ := credentialSlotPresence(w.g.db, name, profile)
-				if len(present) > 0 {
-					ids = append(ids, name)
-				}
-			}
-		}
-		if ids != nil {
-			a.Integrations = ids
-		}
-	}
-	writeJSON(rw, snap)
-}
-
-// apiAgentDelete drops a device from clawpatrol's view. Removes the
-// in-memory agent record, the onboard owner / hostname / profile
-// row, and (in WG mode) the WireGuard peer + allowed-IP entry — so
-// traffic from the deleted device's tunnel can't keep flowing under
-// the old owner. The Tailscale node itself isn't kicked; admins do
-// that from the Tailscale admin console.
-func (w *webMux) apiAgentDelete(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	ip := r.URL.Query().Get("ip")
-	if ip == "" {
-		http.Error(rw, "missing ip", 400)
-		return
-	}
-	if w.g.agents != nil {
-		w.g.agents.Delete(ip)
-	}
-	if w.g.onboard != nil {
-		w.g.onboard.ForgetIP(ip)
-	}
-	if globalWG != nil {
-		globalWG.RevokePeerByIP(ip)
-	}
-	writeJSON(rw, map[string]bool{"ok": true})
-}
-
-// apiAgentProfile assigns a peer IP to a named profile. Profile must
-// be declared in cfg.Profiles. The mapping is persisted in
-// onboard.profileByIP and consulted by Gateway.profileFor at MITM
-// time, so rule scoping switches over immediately.
-func (w *webMux) apiAgentProfile(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	ip := r.URL.Query().Get("ip")
-	profile := r.URL.Query().Get("profile")
-	if ip == "" || profile == "" {
-		http.Error(rw, "missing ip or profile", 400)
-		return
-	}
-	names := orderedProfileNames(w.g.cfg.Policy)
-	known := false
-	for _, n := range names {
-		if n == profile {
-			known = true
-			break
-		}
-	}
-	if !known {
-		http.Error(rw, "unknown profile", 400)
-		return
-	}
-	w.g.onboard.AssignProfile(ip, profile)
-	writeJSON(rw, map[string]any{"ok": true, "ip": ip, "profile": profile})
-}
-
-// apiProfiles lists declared profile names so the dashboard can
-// render a profile picker per device.
-func (w *webMux) apiProfiles(rw http.ResponseWriter, _ *http.Request) {
-	writeJSON(rw, orderedProfileNames(w.g.cfg.Policy))
-}
-
-// RuleSummary is the JSON shape the dashboard renders for each rule.
-// It flattens a CompiledRule plus its enclosing endpoint and profile
-// context so the table view doesn't need to walk the policy graph
-// itself.
-type RuleSummary struct {
-	Name     string                `json:"name"`
-	Family   string                `json:"family"` // "https" | "sql" | "k8s"
-	Endpoint string                `json:"endpoint"`
-	Profile  string                `json:"profile,omitempty"`
-	Priority int                   `json:"priority,omitempty"`
-	Disabled bool                  `json:"disabled,omitempty"`
-	Match    map[string]any        `json:"match,omitempty"`
-	Verdict  string                `json:"verdict,omitempty"`
-	Reason   string                `json:"reason,omitempty"`
-	Approve  []config.ApproveStage `json:"approve,omitempty"`
-}
-
-// apiRules returns every compiled rule across every profile, flattened
-// for the dashboard table view. Rules attached to multiple endpoints
-// emit one row per endpoint so the operator sees each attachment
-// site individually.
-//
-// Read-only. Edits go through PUT /api/config (whole-file HCL via
-// the new typed-block validator).
-func (w *webMux) apiRules(rw http.ResponseWriter, r *http.Request) {
-	switch r.Method {
-	case "GET":
-		writeJSON(rw, w.collectRuleSummaries(""))
-	default:
-		http.Error(rw, "edit rules through PUT /api/config", http.StatusNotImplemented)
-	}
-}
-
-// collectRuleSummaries walks the compiled policy and emits one
-// RuleSummary per (rule × endpoint × profile) triple. When profile is
-// empty, every profile contributes; otherwise only that profile.
-//
-// Sort: by profile, then endpoint, then descending priority (so the
-// dashboard mirrors first-match-wins order within each endpoint).
-func (w *webMux) collectRuleSummaries(profileFilter string) []RuleSummary {
-	policy := w.g.Policy()
-	if policy == nil {
-		return []RuleSummary{}
-	}
-	var out []RuleSummary
-	for profileName, prof := range policy.Profiles {
-		if profileFilter != "" && profileName != profileFilter {
-			continue
-		}
-		for epName, ep := range prof.Endpoints {
-			for _, r := range ep.Rules {
-				out = append(out, RuleSummary{
-					Name:     r.Name,
-					Family:   ep.Family,
-					Endpoint: epName,
-					Profile:  profileName,
-					Priority: r.Priority,
-					Disabled: r.Disabled,
-					Match:    matchSourceMap(r),
-					Verdict:  r.Outcome.Verdict,
-					Reason:   r.Outcome.Reason,
-					Approve:  r.Outcome.Approve,
-				})
-			}
-		}
-	}
-	sort.Slice(out, func(i, j int) bool {
-		if out[i].Profile != out[j].Profile {
-			return out[i].Profile < out[j].Profile
-		}
-		if out[i].Endpoint != out[j].Endpoint {
-			return out[i].Endpoint < out[j].Endpoint
-		}
-		return out[i].Priority > out[j].Priority
-	})
-	return out
-}
-
-func matchSourceMap(r *config.CompiledRule) map[string]any {
-	if r == nil {
-		return nil
-	}
-	return r.Match
-}
-
 // apiConfig serves the entire gateway.hcl for the global settings
 // editor. GET returns the file as-is (preserves operator comments).
 // PUT validates by re-parsing + writing through writeConfigHCL so
@@ -977,320 +646,6 @@ func isLoopback(host string) bool {
 	return host == "127.0.0.1" || host == "::1" || strings.HasPrefix(host, "127.")
 }
 
-func (w *webMux) apiOAuthStart(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	id := r.URL.Query().Get("id")
-	flow := lookupOAuthFlow(w.g.policy.Load(), id)
-	if flow == nil {
-		http.Error(rw, "no oauth integration: "+id, 400)
-		return
-	}
-	owner, ownerLabel := w.ownerForCaller(r)
-	if owner == "" {
-		http.Error(rw, "could not determine owner identity (tailscale whois failed)", 400)
-		return
-	}
-	// Branch: device flow vs auth-code+PKCE.
-	if flow.Flow == "device" {
-		w.startDeviceFlow(rw, id, flow, owner, ownerLabel)
-		return
-	}
-
-	verifier := randomString(64)
-	sum := sha256.Sum256([]byte(verifier))
-	challenge := base64.RawURLEncoding.EncodeToString(sum[:])
-	state := randomString(32)
-	cfg := &oauth2.Config{
-		ClientID:     resolveTemplate(flow.OAuth.ClientID),
-		ClientSecret: resolveTemplate(flow.OAuth.ClientSecret),
-		Scopes:       flow.OAuth.Scopes,
-		RedirectURL:  flow.OAuth.RedirectURI,
-		Endpoint:     oauth2.Endpoint{AuthURL: flow.OAuth.AuthURL, TokenURL: flow.OAuth.TokenURL},
-	}
-	authURL := cfg.AuthCodeURL(state,
-		oauth2.SetAuthURLParam("code_challenge", challenge),
-		oauth2.SetAuthURLParam("code_challenge_method", "S256"),
-	)
-	w.mu.Lock()
-	w.sessions[state] = &oauthSession{verifier: verifier, state: state, cfg: cfg, id: id, owner: owner, created: time.Now()}
-	for k, s := range w.sessions {
-		if time.Since(s.created) > 10*time.Minute {
-			delete(w.sessions, k)
-		}
-	}
-	w.mu.Unlock()
-	writeJSON(rw, map[string]string{"auth_url": authURL, "state": state, "owner": ownerLabel})
-}
-
-func (w *webMux) apiOAuthExchange(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	var body struct {
-		State string `json:"state"`
-		Code  string `json:"code"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		http.Error(rw, err.Error(), 400)
-		return
-	}
-	body.Code = strings.TrimSpace(body.Code)
-	if body.Code == "" || body.State == "" {
-		http.Error(rw, "missing code/state", 400)
-		return
-	}
-	if i := strings.IndexAny(body.Code, "#&?"); i > 0 {
-		body.Code = body.Code[:i]
-	}
-
-	w.mu.Lock()
-	sess, ok := w.sessions[body.State]
-	if ok {
-		delete(w.sessions, body.State)
-	}
-	w.mu.Unlock()
-	if !ok {
-		http.Error(rw, "unknown state (expired or stale)", 400)
-		return
-	}
-
-	ctx, cancel := context.WithTimeout(r.Context(), 30*time.Second)
-	defer cancel()
-	tok, err := exchangeOAuthCode(ctx, sess, body.Code, body.State)
-	if err != nil {
-		http.Error(rw, "token exchange: "+err.Error(), 400)
-		return
-	}
-	if err := w.g.oauth.Set(sess.id, sess.owner, tok); err != nil {
-		http.Error(rw, err.Error(), 500)
-		return
-	}
-	writeJSON(rw, map[string]any{"connected": true, "owner": sess.owner, "expires": tok.Expiry.Unix()})
-}
-
-// startDeviceFlow kicks off OAuth device flow (RFC 8628). Returns
-// {user_code, verification_uri, device_code, interval} so the dashboard
-// can prompt the user to enter the code at the verification URI.
-func (w *webMux) startDeviceFlow(rw http.ResponseWriter, id string, it *OAuthIntegration, owner, ownerLabel string) {
-	form := url.Values{}
-	form.Set("client_id", resolveTemplate(it.OAuth.ClientID))
-	if len(it.OAuth.Scopes) > 0 {
-		form.Set("scope", strings.Join(it.OAuth.Scopes, " "))
-	}
-	req, _ := http.NewRequest("POST", it.OAuth.DeviceURL, strings.NewReader(form.Encode()))
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Set("Accept", "application/json")
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		http.Error(rw, "device-code: "+err.Error(), 502)
-		return
-	}
-	defer resp.Body.Close()
-	body, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode != 200 {
-		http.Error(rw, fmt.Sprintf("device-code %d: %s", resp.StatusCode, string(body)), 502)
-		return
-	}
-	var dr struct {
-		DeviceCode      string `json:"device_code"`
-		UserCode        string `json:"user_code"`
-		VerificationURI string `json:"verification_uri"`
-		ExpiresIn       int    `json:"expires_in"`
-		Interval        int    `json:"interval"`
-	}
-	if err := json.Unmarshal(body, &dr); err != nil {
-		http.Error(rw, "device-code parse: "+err.Error(), 502)
-		return
-	}
-	state := randomString(32)
-	w.mu.Lock()
-	w.sessions[state] = &oauthSession{
-		state:    state,
-		id:       id,
-		owner:    owner,
-		created:  time.Now(),
-		verifier: dr.DeviceCode, // reuse field for device_code
-		cfg: &oauth2.Config{
-			ClientID: resolveTemplate(it.OAuth.ClientID),
-			Endpoint: oauth2.Endpoint{TokenURL: it.OAuth.TokenURL},
-		},
-	}
-	w.mu.Unlock()
-	writeJSON(rw, map[string]any{
-		"flow":             "device",
-		"state":            state,
-		"user_code":        dr.UserCode,
-		"verification_uri": dr.VerificationURI,
-		"interval":         dr.Interval,
-		"expires_in":       dr.ExpiresIn,
-		"owner":            ownerLabel,
-	})
-}
-
-// pollDeviceFlow exchanges device_code for a token. Called by the
-// frontend on a timer until success / denial / expiration.
-func (w *webMux) pollDeviceFlow(rw http.ResponseWriter, sess *oauthSession) {
-	form := url.Values{}
-	form.Set("client_id", sess.cfg.ClientID)
-	form.Set("device_code", sess.verifier)
-	form.Set("grant_type", "urn:ietf:params:oauth:grant-type:device_code")
-	req, _ := http.NewRequest("POST", sess.cfg.Endpoint.TokenURL, strings.NewReader(form.Encode()))
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	req.Header.Set("Accept", "application/json")
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		http.Error(rw, "device poll: "+err.Error(), 502)
-		return
-	}
-	defer resp.Body.Close()
-	body, _ := io.ReadAll(resp.Body)
-	// Don't log the body verbatim — on success it carries access_token.
-	var tr struct {
-		AccessToken      string `json:"access_token"`
-		TokenType        string `json:"token_type"`
-		RefreshToken     string `json:"refresh_token"`
-		Scope            string `json:"scope"`
-		ExpiresIn        int64  `json:"expires_in"`
-		Error            string `json:"error"`
-		ErrorDescription string `json:"error_description"`
-		Interval         int    `json:"interval"`
-	}
-	if err := json.Unmarshal(body, &tr); err != nil {
-		http.Error(rw, "device poll parse: "+err.Error(), 502)
-		return
-	}
-	if tr.Error != "" {
-		// `slow_down` carries an updated interval (RFC 8628). Surface
-		// it to the dashboard so the polling loop respects the new
-		// cadence; otherwise the client keeps hitting at the original
-		// interval and GitHub never returns the token.
-		out := map[string]any{"error": tr.Error, "detail": tr.ErrorDescription}
-		if tr.Interval > 0 {
-			out["interval"] = tr.Interval
-		}
-		writeJSON(rw, out)
-		return
-	}
-	if tr.AccessToken == "" {
-		writeJSON(rw, map[string]string{"error": "authorization_pending"})
-		return
-	}
-	tok := &oauth2.Token{
-		AccessToken:  tr.AccessToken,
-		RefreshToken: tr.RefreshToken,
-		TokenType:    tr.TokenType,
-	}
-	if tr.ExpiresIn > 0 {
-		tok.Expiry = time.Now().Add(time.Duration(tr.ExpiresIn) * time.Second)
-	}
-	w.mu.Lock()
-	delete(w.sessions, sess.state)
-	w.mu.Unlock()
-	if err := w.g.oauth.Set(sess.id, sess.owner, tok); err != nil {
-		http.Error(rw, err.Error(), 500)
-		return
-	}
-	writeJSON(rw, map[string]any{"connected": true, "owner": sess.owner})
-}
-
-// exchangeOAuthCode finalizes an OAuth PKCE flow. Anthropic's token
-// endpoint requires a JSON body (returns "Invalid request format" for
-// the standard form-urlencoded body), so we hand-roll the request for
-// claude integration. Other providers use the stdlib oauth2.Exchange.
-func exchangeOAuthCode(ctx context.Context, sess *oauthSession, code, state string) (*oauth2.Token, error) {
-	if sess.id == "claude" {
-		return exchangeAnthropicCode(ctx, sess, code, state)
-	}
-	return sess.cfg.Exchange(ctx, code,
-		oauth2.SetAuthURLParam("code_verifier", sess.verifier),
-		oauth2.SetAuthURLParam("redirect_uri", sess.cfg.RedirectURL),
-	)
-}
-
-func exchangeAnthropicCode(ctx context.Context, sess *oauthSession, code, state string) (*oauth2.Token, error) {
-	body, _ := json.Marshal(map[string]string{
-		"grant_type":    "authorization_code",
-		"code":          code,
-		"redirect_uri":  sess.cfg.RedirectURL,
-		"client_id":     sess.cfg.ClientID,
-		"code_verifier": sess.verifier,
-		"state":         state,
-	})
-	req, err := http.NewRequestWithContext(ctx, "POST", sess.cfg.Endpoint.TokenURL, strings.NewReader(string(body)))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	respBytes, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode != 200 {
-		return nil, fmt.Errorf("anthropic %d: %s", resp.StatusCode, string(respBytes))
-	}
-	var tr struct {
-		AccessToken  string `json:"access_token"`
-		RefreshToken string `json:"refresh_token"`
-		TokenType    string `json:"token_type"`
-		ExpiresIn    int64  `json:"expires_in"`
-	}
-	if err := json.Unmarshal(respBytes, &tr); err != nil {
-		return nil, err
-	}
-	tok := &oauth2.Token{
-		AccessToken:  tr.AccessToken,
-		RefreshToken: tr.RefreshToken,
-		TokenType:    tr.TokenType,
-	}
-	if tr.ExpiresIn > 0 {
-		tok.Expiry = time.Now().Add(time.Duration(tr.ExpiresIn) * time.Second)
-	}
-	return tok, nil
-}
-
-func (w *webMux) apiOAuthDevicePoll(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	state := r.URL.Query().Get("state")
-	w.mu.Lock()
-	sess := w.sessions[state]
-	w.mu.Unlock()
-	if sess == nil {
-		http.Error(rw, "unknown state (expired)", 400)
-		return
-	}
-	w.pollDeviceFlow(rw, sess)
-}
-
-func (w *webMux) apiOAuthRevoke(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	var body struct {
-		ID    string `json:"id"`
-		Owner string `json:"owner"`
-	}
-	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
-		http.Error(rw, err.Error(), 400)
-		return
-	}
-	if body.ID == "" || body.Owner == "" {
-		http.Error(rw, "missing id or owner", 400)
-		return
-	}
-	w.g.oauth.Revoke(body.ID, body.Owner)
-	writeJSON(rw, map[string]bool{"ok": true})
-}
-
 func (w *webMux) apiEventsSSE(rw http.ResponseWriter, r *http.Request) {
 	rw.Header().Set("Content-Type", "text/event-stream")
 	rw.Header().Set("Cache-Control", "no-cache")
@@ -1365,208 +720,6 @@ func writeJSON(rw http.ResponseWriter, v any) {
 // since this IS how a brand-new client first contacts the gateway.
 // The returned user_code must still be approved by an existing tailnet
 // member on the dashboard.
-func (w *webMux) apiOnboardStart(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	s := w.onboard.start()
-	// CLI passes its os.Hostname() so the dashboard shows a real
-	// device name instead of just the WG-side IP. Optional — we still
-	// fall back gracefully when missing.
-	if hn := strings.TrimSpace(r.URL.Query().Get("hostname")); hn != "" {
-		w.onboard.mu.Lock()
-		s.hostname = hn
-		w.onboard.mu.Unlock()
-	}
-	// Prefer the operator-configured public_url so brand-new clients
-	// see a real URL. Fall back to the request's Host header for
-	// dev / direct-IP setups. Tailscale Funnel hostnames (*.ts.net)
-	// are always HTTPS, so detect those explicitly.
-	verifyURL := w.publicURL
-	if verifyURL == "" {
-		scheme := "http"
-		if strings.HasSuffix(r.Host, ".ts.net") || r.TLS != nil || r.Header.Get("X-Forwarded-Proto") == "https" {
-			scheme = "https"
-		}
-		verifyURL = scheme + "://" + r.Host
-	}
-	verifyURL = strings.TrimRight(verifyURL, "/") + "/#/onboard/" + s.userCode
-	writeJSON(rw, map[string]any{
-		"device_code": s.deviceCode,
-		"user_code":   s.userCode,
-		"verify_url":  verifyURL,
-		"interval":    3,
-		"expires_in":  600,
-	})
-}
-
-// apiOnboardLookup returns user_code session info for the dashboard
-// approval page. No secrets exposed (just code + age).
-func (w *webMux) apiOnboardLookup(rw http.ResponseWriter, r *http.Request) {
-	code := r.URL.Query().Get("code")
-	s := w.onboard.byUserCode(code)
-	if s == nil {
-		http.Error(rw, "unknown or expired code", 404)
-		return
-	}
-	writeJSON(rw, map[string]any{
-		"user_code":  s.userCode,
-		"approved":   s.approved,
-		"created_at": s.created.Unix(),
-	})
-}
-
-// apiOnboardApprove is hit by the dashboard "approve" button. The
-// caller must be an existing tailnet member (whois succeeds) — this
-// gates onboarding behind an existing trusted user.
-func (w *webMux) apiOnboardApprove(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	owner, _ := w.ownerForCaller(r)
-	if owner == "" {
-		http.Error(rw, "approval requires an authenticated tailnet caller (or set admin_email in gateway.hcl)", 403)
-		return
-	}
-	code := r.URL.Query().Get("code")
-	// Operator picks which profile this device joins. Falls back to a
-	// profile named "default" when declared; otherwise the first profile
-	// in source order.
-	profile := r.URL.Query().Get("profile")
-	if profile == "" {
-		profile = defaultProfileName(w.g.cfg.Policy)
-	}
-	s := w.onboard.byUserCode(code)
-	if s == nil {
-		http.Error(rw, "unknown or expired code", 404)
-		return
-	}
-	w.onboard.mu.Lock()
-	if s.approved {
-		w.onboard.mu.Unlock()
-		writeJSON(rw, map[string]any{"already": true})
-		return
-	}
-	s.approved = true
-	s.owner = owner
-	s.profile = profile
-	hostname := s.hostname
-	w.onboard.mu.Unlock()
-
-	// Recycle the wg /32 already bound to (owner, hostname) so a rejoin
-	// from the same machine doesn't spawn a duplicate device row.
-	reuseIP := w.onboard.IPForHostname(owner, hostname)
-
-	// Mint key in background so the approve click returns fast.
-	go func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-		key, loginServer, peerIP, err := newOnboarder(w.ts).MintKey(ctx, reuseIP)
-		w.onboard.mu.Lock()
-		if err != nil {
-			s.err = err.Error()
-			w.onboard.mu.Unlock()
-			return
-		}
-		s.authKey = key
-		s.loginServer = loginServer
-		dc := s.deviceCode
-		w.onboard.mu.Unlock()
-		// WG path: server allocated peerIP and knows the approver, so
-		// register the (ip → owner) mapping right here. Saves the CLI
-		// from making a /api/onboard/claim round-trip after wg-quick
-		// (which is racy: the default route is now the tunnel and the
-		// public gateway URL becomes unreachable).
-		if peerIP != "" {
-			w.onboard.ClaimIP(dc, peerIP, "")
-			if profile != "" {
-				w.onboard.AssignProfile(peerIP, profile)
-			}
-			// Seed the agents registry so the dashboard shows the
-			// device immediately, before it sends any traffic.
-			if w.g.agents != nil {
-				w.g.agents.Seed(peerIP)
-			}
-		}
-	}()
-	writeJSON(rw, map[string]any{"approved": true})
-}
-
-// apiOnboardClaim is hit by the CLI right after `tailscale up`
-// finishes. The peer IP (the new device's tailnet IP) gets associated
-// with the approver email. Subsequent agent traffic from that IP
-// resolves to the approver in lieu of the useless "tagged-devices"
-// whois result.
-func (w *webMux) apiOnboardClaim(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	dc := r.URL.Query().Get("device_code")
-	if dc == "" {
-		http.Error(rw, "missing device_code", 400)
-		return
-	}
-	// CLI passes its own tailnet IP via ?ip=. We can't trust
-	// r.RemoteAddr because tailscale funnel proxies through localhost
-	// and the device_code already binds this to a specific approval —
-	// you can't claim someone else's IP without their device_code.
-	host := r.URL.Query().Get("ip")
-	if host == "" {
-		host = r.RemoteAddr
-		if i := strings.LastIndex(host, ":"); i >= 0 {
-			host = host[:i]
-		}
-	}
-	hostname := strings.TrimSpace(r.URL.Query().Get("hostname"))
-	owner, ok := w.onboard.ClaimIP(dc, host, hostname)
-	if !ok {
-		log.Printf("onboard claim: unknown/unapproved device_code=%s ip=%s", truncate(dc, 16), host)
-		http.Error(rw, "unknown or unapproved device_code", 404)
-		return
-	}
-	log.Printf("onboard claim: %s → %s (hostname=%q)", host, owner, hostname)
-	writeJSON(rw, map[string]string{"owner": owner, "ip": host})
-}
-
-// apiOnboardPoll is hit by the CLI to retrieve the auth key once
-// approved. Uses standard device-flow status codes via JSON.
-func (w *webMux) apiOnboardPoll(rw http.ResponseWriter, r *http.Request) {
-	if r.Method != "POST" {
-		http.Error(rw, "POST", 405)
-		return
-	}
-	dc := r.URL.Query().Get("device_code")
-	s := w.onboard.byDeviceCode(dc)
-	if s == nil {
-		writeJSON(rw, map[string]string{"error": "expired_token"})
-		return
-	}
-	w.onboard.mu.Lock()
-	defer w.onboard.mu.Unlock()
-	if s.err != "" {
-		writeJSON(rw, map[string]string{"error": "server_error", "detail": s.err})
-		return
-	}
-	if !s.approved {
-		writeJSON(rw, map[string]string{"error": "authorization_pending"})
-		return
-	}
-	if s.authKey == "" {
-		writeJSON(rw, map[string]string{"error": "slow_down"})
-		return
-	}
-	writeJSON(rw, map[string]any{
-		"auth_key":     s.authKey,
-		"approved_by":  s.owner,
-		"login_server": s.loginServer, // empty = Tailscale Inc default
-	})
-}
-
-var displayOrder = []string{"claude", "codex", "github"}
-
 func allIntegrationKeys() []string { return displayOrder }
 
 // Event sink + sampling helpers (fed by g.handle/mitm/splice; consumed
